Change refactor enet clients to have an abtract base class with one client using threads, the other one using tasks
